+++ src/datagarden_models/models/economics/__init__.py
@@ -4,16 +4,20 @@ from ..base import DataGardenModel, DataGardenModelLegends
 from .base_economics import EconomicBaseKeys
 from .gdp import GDP, GDPV1Keys
 from .infllation import Inflation, InflationV1Keys
+from .trade import TradeV1, TradeV1Keys
 
 
-class EconomicsV1Keys(GDPV1Keys, EconomicBaseKeys, InflationV1Keys):
+class EconomicsV1Keys(GDPV1Keys, EconomicBaseKeys, InflationV1Keys, TradeV1Keys):
 	GDP = "gdp"
 	DATAGARDEN_MODEL_NAME = "Economics"
+	INFLATION = "inflation"
+	TRADE = "trade"
 
 
 class EconomicsV1Legends(DataGardenModelLegends):
 	GDP = "Gross Domestic Product"
-	INFLATION = "Inflation"
+	INFLATION = "Inflation numbers"
+	TRADE = "Trade statistics"
 
 
 L = EconomicsV1Legends
@@ -26,3 +30,4 @@ class EconomicsV1(DataGardenModel):
 	)
 	gdp: GDP = Field(default_factory=GDP, description=L.GDP)
 	inflation: Inflation = Field(default_factory=Inflation, description=L.INFLATION)
+	trade: TradeV1 = Field(default_factory=TradeV1, description=L.TRADE)